+++ parser_library/src/expressions/conditional_assembly/ca_operator_binary.h
@@ -31,8 +31,7 @@ public:
 
     undef_sym_set get_undefined_attributed_symbols(const evaluation_context& eval_ctx) const override;
 
-    void resolve_expression_tree(
-        context::SET_t_enum kind, context::SET_t_enum parent_expr_kind, diagnostic_op_consumer& diags) override;
+    void resolve_expression_tree(ca_expression_ctx expr_ctx, diagnostic_op_consumer& diags) override;
 
     bool is_character_expression(character_expression_purpose purpose) const override;
 
@@ -70,8 +69,7 @@ public:
         range expr_range,
         context::SET_t_enum parent_expr_kind = context::SET_t_enum::UNDEF_TYPE);
 
-    void resolve_expression_tree(
-        context::SET_t_enum kind, context::SET_t_enum parent_expr_kind, diagnostic_op_consumer& diags) override;
+    void resolve_expression_tree(ca_expression_ctx expr_ctx, diagnostic_op_consumer& diags) override;
 
     context::SET_t operation(context::SET_t lhs, context::SET_t rhs, const evaluation_context& eval_ctx) const override;
 
@@ -81,7 +79,7 @@ public:
 private:
     bool is_relational() const;
     ca_expr_ops function;
-    context::SET_t_enum m_parent_expr_kind;
+    ca_expression_ctx m_expr_ctx;
 };
 
 struct ca_add