+++ parser_library/src/expressions/conditional_assembly/terms/ca_function.cpp
@@ -14,6 +14,7 @@
 
 #include "ca_function.h"
 
+#include <array>
 #include <cassert>
 #include <charconv>
 #include <iomanip>
@@ -56,22 +57,29 @@ undef_sym_set ca_function::get_undefined_attributed_symbols(const evaluation_con
     return ret;
 }
 
-void ca_function::resolve_expression_tree(
-    context::SET_t_enum kind, context::SET_t_enum parent_expr_kind, diagnostic_op_consumer& diags)
+void ca_function::resolve_expression_tree(ca_expression_ctx expr_ctx, diagnostic_op_consumer& diags)
 {
-    if (duplication_factor && expr_kind != context::SET_t_enum::C_TYPE)
+    // No diag when kind == expr_kind or when there is a combination of A_TYPE and B_TYPE
+    static constexpr bool allowed_combinations[4][4] = {
+        { 1, 1, 0, 0 },
+        { 1, 1, 0, 0 },
+        { 0, 0, 1, 0 },
+        { 0, 0, 0, 1 },
+    };
+
+    if (!allowed_combinations[static_cast<int>(expr_ctx.kind)][static_cast<int>(expr_kind)])
+        diags.add_diagnostic(diagnostic_op::error_CE004(expr_range));
+    else if (duplication_factor && expr_kind != context::SET_t_enum::C_TYPE)
         diags.add_diagnostic(diagnostic_op::error_CE005(duplication_factor->expr_range));
+    else if (auto [param_size, param_kind] = ca_common_expr_policy::get_function_param_info(function, expr_kind);
+             parameters.size() != param_size)
+        diags.add_diagnostic(diagnostic_op::error_CE006(expr_range));
     else
     {
-        auto [param_size, param_kind] = ca_common_expr_policy::get_function_param_info(function, expr_kind);
-        if (parameters.size() != param_size)
-            diags.add_diagnostic(diagnostic_op::error_CE006(expr_range));
-        else
+        expr_ctx.kind = param_kind;
+        for (auto&& expr : parameters)
         {
-            for (auto&& expr : parameters)
-            {
-                expr->resolve_expression_tree(param_kind, parent_expr_kind, diags);
-            }
+            expr->resolve_expression_tree(expr_ctx, diags);
         }
     }
 }