+++ parser_library/src/expressions/conditional_assembly/terms/ca_var_sym.cpp
@@ -48,13 +48,13 @@ undef_sym_set ca_var_sym::get_undefined_attributed_symbols(const evaluation_cont
     return get_undefined_attributed_symbols_vs(symbol, eval_ctx);
 }
 
-void ca_var_sym::resolve_expression_tree(context::SET_t_enum kind, context::SET_t_enum, diagnostic_op_consumer& diags)
+void ca_var_sym::resolve_expression_tree(ca_expression_ctx expr_ctx, diagnostic_op_consumer& diags)
 {
     // this conversion request indicates that the variable was used without the mandatory quotes around it
-    if (kind == context::SET_t_enum::C_TYPE)
+    if (expr_ctx.kind == context::SET_t_enum::C_TYPE)
         diags.add_diagnostic(diagnostic_op::error_CE017_character_expression_expected(expr_range));
-    expr_kind = kind;
-    symbol->resolve(diags);
+    expr_kind = expr_ctx.kind;
+    symbol->resolve(expr_ctx.parent_expr_kind, diags);
 }
 
 bool ca_var_sym::is_character_expression(character_expression_purpose) const { return false; }