+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 - Implement rest of the changes introduced by APAR PH46868
+- Evaluation of subscripted expressions in CA statements
 
 ## [1.4.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.3.0...1.4.0) (2022-08-30)